requirements.txt
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
faiss-cpu>=1.7.4
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
